requirements.txt
langchain-openai>=0.3.0
langgraph>=0.3.0
langchain-core>=0.3.0
kagent-langgraph>=0.1.0
httpx>=0.27.0
uvicorn>=0.34.0
fastapi>=0.115.0
